modularized code and added tests
